LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-on-focus-expected.txt


FAIL scroll snap should happens on Element.focus() assert_equals: expected 1200 but got 0
